frontend: separate list to list component
