+++ frontend/Component/Floating.tsx
@@ -3,6 +3,7 @@ import type { FeatureCollection } from "geojson";
 import "./Floating.css";
 import axios from "axios";
 import Filter from "./Filter";
+import { List } from "./List";
 
 type UploadStatus = "idle" | "uploading" | "success" | "error";
 
@@ -40,13 +41,6 @@ interface FloatingProps {
   handleGeoJSON?: (fc: any) => void;
 }
 
-function shortenName(name: string, maxLength = 25) {
-  if (name.length <= maxLength) return name;
-  const start = name.slice(0, 10); // first part
-  const end = name.slice(-10); // last part
-  return `${start}...${end}`;
-}
-
 export default function Floating({ handleGeoJSON }: FloatingProps) {
   /* File Data */
   const [file, setFile] = useState<File | null>(null);
@@ -64,10 +58,6 @@ export default function Floating({ handleGeoJSON }: FloatingProps) {
   const [openLayerChildren, setOpenLayerChildren] = useState<Children[] | null>(
     null,
   );
-
-  /** Filter Select */
-  const [selectedCity, setSelectedCity] = useState<string | null>(null);
-  const [selectedRegion, setSelectedRegion] = useState<string | null>(null);
 
   const [openDropdown, setOpenDropdown] = useState<"region" | "city" | null>(
     null,
@@ -137,15 +127,6 @@ export default function Floating({ handleGeoJSON }: FloatingProps) {
       }
       setFile(selectedFile);
       setFormError(null);
-    }
-  }
-
-  async function handleDelete(fileId: string) {
-    try {
-      await axios.delete(`http://localhost:3000/files/${fileId}`);
-      getKMZList();
-    } catch (error) {
-      console.log(error);
     }
   }
 
@@ -361,103 +342,14 @@ export default function Floating({ handleGeoJSON }: FloatingProps) {
         </div>
 
         {/** Separate Component */}
-        <div className="List">
-          <h4> Stored Queries </h4>
-          <ul className="List-wrapper">
-            {layerList && layerList.length > 0
-              ? layerList.map((l) => {
-                  const isOpen = openLayer === l.id;
-                  return (
-                    <>
-                      <li
-                        className={`List-item ${isOpen ? "is-open" : ""}`}
-                        key={l.id}
-                        role="button"
-                        aria-expanded={isOpen}
-                        aria-controls={`panel-${l.id}`}
-                        onKeyDown={(e) =>
-                          (e.key === "Enter" || e.key === " ") &&
-                          updateMapView(l.id)
-                        }
-                      >
-                        <div className="List-data">
-                          <p>
-                            <span
-                              className={`caret ${isOpen ? "down" : ""}`}
-                              aria-hidden
-                            />
-                            {shortenName(l.name)}
-                          </p>
-
-                          <div className="List-actions">
-                            <button
-                              className="List-view"
-                              onClick={(e) => {
-                                e.stopPropagation();
-                                updateMapView(l.id);
-                                // toggleOpen(l.id); // Handle TOggle
-                              }}
-                            >
-                              View
-                            </button>
-                            <button
-                              className="List-delete"
-                              onClick={(e) => {
-                                e.stopPropagation();
-                                handleDelete(l.id);
-                              }}
-                            >
-                              Delete
-                            </button>
-                          </div>
-                        </div>
-
-                        {/* DROP DOWN */}
-                        {isOpen && (
-                          <div className="List-dropdown" id={`panel-${l.id}`}>
-                            {openLayerChildren
-                              ? openLayerChildren.map((child) => {
-                                  return (
-                                    <>
-                                      <div className="Content">
-                                        <input
-                                          type="checkbox"
-                                          id={`child-${child.id}`}
-                                          checked={child.isChecked}
-                                          onChange={(e) => {
-                                            console.log(
-                                              `${child.name} toggled:`,
-                                              e.target.checked,
-                                            );
-                                            e.stopPropagation();
-                                            toggleChildCheckbox(
-                                              child.id,
-                                              e.target.checked,
-                                            );
-                                          }}
-                                          style={{
-                                            cursor: "pointer",
-                                            margin: "0 10px",
-                                          }}
-                                        />
-                                        <label htmlFor={`child-${child.id}`}>
-                                          {child.name}
-                                        </label>
-                                      </div>
-                                    </>
-                                  );
-                                })
-                              : null}
-                          </div>
-                        )}
-                      </li>
-                    </>
-                  );
-                })
-              : null}
-          </ul>
-        </div>
-
+        <List
+          updateMapView={updateMapView}
+          getKMZList={getKMZList}
+          toggleChildCheckbox={toggleChildCheckbox}
+          openLayerChildren={openLayerChildren}
+          layerList={layerList}
+          openLayer={openLayer}
+        />
         <Filter REGION_OPTIONS={REGION_OPTIONS} CITY_OPTIONS={CITY_OPTIONS} />
       </div>
     </>
